:truck: Moved uncertainty visualisations into plotting_core and plotting_static
