+++ icenet_mp/visualisations/__init__.py
@@ -4,7 +4,8 @@ from matplotlib import rcParams
 from .dataset_plotting import plot_variables_static, plot_variables_video
 from .helpers import DEFAULT_SIC_SPEC
 from .plotter import Plotter
-from .uncertainty import compute_standardised_difference, plot_static_uncertainty
+from .plotting_core import compute_standardised_difference
+from .plotting_static import plot_static_uncertainty
 
 
 def register_animation_backends() -> None:
